feat: show react sdk to people
