+++ client-sdks/react/package.json
@@ -1,17 +1,29 @@
 {
-  "name": "@mastra/react-hooks",
-  "version": "0.0.1-alpha.1",
-  "private": true,
+  "name": "@mastra/react",
+  "version": "0.0.1-alpha.0",
+  "private": false,
   "type": "module",
   "main": "dist/index.cjs.js",
   "module": "dist/index.es.js",
   "types": "dist/index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/mastra-ai/mastra.git",
+    "directory": "client-sdks/react"
+  },
+  "bugs": {
+    "url": "https://github.com/mastra-ai/mastra/issues"
+  },
+  "homepage": "https://mastra.ai",
+  "license": "Apache-2.0",
   "scripts": {
     "build": "vite build",
     "dev": "vite build --watch"
   },
   "files": [
-    "src"
+    "dist",
+    "CHANGELOG.md",
+    "README.md"
   ],
   "dependencies": {
     "@mastra/client-js": "workspace:*",
@@ -20,14 +32,15 @@
   },
   "devDependencies": {
     "@assistant-ui/react": "^0.10.45",
-    "eslint": "^9.35.0",
-    "typescript": "^5.8.3",
+    "@mastra/core": "workspace:*",
     "@types/react": "^19.1.13",
     "@types/react-dom": "^19.1.9",
-    "@mastra/core": "workspace:*",
-    "vite": "^7.1.7",
-    "vite-plugin-dts": "^4.5.4",
+    "@vitejs/plugin-react": "^5.0.3",
+    "ai": "^5.0.52",
+    "eslint": "^9.35.0",
     "rollup-plugin-node-externals": "^8.0.1",
-    "@vitejs/plugin-react": "^5.0.3"
+    "typescript": "^5.8.3",
+    "vite": "^7.1.7",
+    "vite-plugin-dts": "^4.5.4"
   }
 }
