+++ client-sdks/react/src/index.ts
@@ -2,3 +2,7 @@ export * from './mastra-react-provider';
 export * from './agent/hooks'; // Agent hooks
 export * from './agent/types';
 export { useMastraClient } from './mastra-client-context';
+
+// Third party integrations
+export * from './agent/lib/ai-sdk/toUIMessage';
+export * from './agent/lib/assistant-ui';
