+++ packages/playground-ui/package.json
@@ -45,7 +45,7 @@
   "author": "Mastra",
   "license": "Apache-2.0",
   "dependencies": {
-    "@mastra/react-hooks": "workspace:*",
+    "@mastra/react": "workspace:*",
     "@ai-sdk/ui-utils": "^1.2.11",
     "@assistant-ui/react": "^0.10.45",
     "@assistant-ui/react-markdown": "^0.10.9",
