+++ packages/playground-ui/src/domains/agents/hooks/use-agent-working-memory.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect, useCallback } from 'react';
-import { useMastraClient } from '@mastra/react-hooks';
+import { useMastraClient } from '@mastra/react';
 
 function parseJsonString(jsonString: string): any {
   try {
