+++ packages/playground-ui/src/domains/voice/hooks/use-speech-recognition.ts
@@ -1,4 +1,4 @@
-import { useMastraClient } from '@mastra/react-hooks';
+import { useMastraClient } from '@mastra/react';
 import { Agent } from '@mastra/core/agent';
 import { useEffect, useRef, useState } from 'react';
 import { recordMicrophoneToFile } from '../utils/record-mic-to-file';
