+++ packages/playground-ui/src/hooks/use-workflow-runs.ts
@@ -1,4 +1,4 @@
-import { useMastraClient } from '@mastra/react-hooks';
+import { useMastraClient } from '@mastra/react';
 import { useQuery } from '@tanstack/react-query';
 
 export const useWorkflowRuns = (workflowId: string, { enabled = true }: { enabled?: boolean } = {}) => {
