+++ packages/playground/package.json
@@ -26,7 +26,7 @@
     "@lukeed/uuid": "^2.0.1",
     "@mastra/client-js": "workspace:*",
     "@mastra/core": "workspace:*",
-    "@mastra/react-hooks": "workspace:*",
+    "@mastra/react": "workspace:*",
     "@mastra/playground-ui": "workspace:^",
     "@radix-ui/react-checkbox": "^1.3.2",
     "@radix-ui/react-dialog": "^1.1.14",
