+++ packages/playground/src/domains/agents/hooks/use-prompt-versions.ts
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 
 import type { EvalResult, PromptVersion } from '../types';
 import { usePlaygroundStore } from '@mastra/playground-ui';
-import { useMastraClient } from '@mastra/react-hooks';
+import { useMastraClient } from '@mastra/react';
 
 export function usePromptVersions(agentId: string, instructions?: string) {
   const [versions, setVersions] = useState<PromptVersion[]>([]);
