+++ packages/playground/src/hooks/use-memory.ts
@@ -2,7 +2,7 @@ import { toast } from 'sonner';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
 import type { MemorySearchResponse, MemorySearchParams } from '@/types/memory';
-import { useMastraClient } from '@mastra/react-hooks';
+import { useMastraClient } from '@mastra/react';
 
 export const useMemory = (agentId?: string) => {
   const client = useMastraClient();
